+++ explorer/assistant/tests.py
@@ -97,7 +97,6 @@ class TestAssistantUtils(TestCase):
         SimpleQueryFactory(title="First Query")
         SimpleQueryFactory(title="Second Query")
         ret = sample_rows_from_tables(CONN, ["explorer_query"])
-        print(ret)
         self.assertTrue("First Query" in ret)
         self.assertTrue("Second Query" in ret)